requirements.txt
amqp==5.1.1
apturl==0.5.2
argh==0.26.2
asgiref==3.7.2
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
blis==0.7.8
boto3==1.21.46
botocore==1.24.46
Brlapi==0.8.3
Brotli==1.0.9
bs4==0.0.1
cachetools==5.3.1
catalogue==2.0.8
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.2.0
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
command-not-found==0.3
comtypes==1.1.14
confection==0.0.3
constantly==15.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
cssselect==1.2.0
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.6
daphne==4.0.0
dateparser==1.1.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dj-database-url==0.5.0
Django==4.2.4
django-allauth==0.51.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-braces==1.15.0
django-celery-email==3.0.0
django-channels==0.7.0
django-ckeditor==6.7.0
django-colorfield==0.7.2
django-compressor==4.1
django-cors-headers==4.2.0
django-countries==7.4.2
django-crispy-forms==1.14.0
django-cursor-pagination==0.2.0
django-cuser==2017.3.16
django-dotenv==1.4.2
django-environ==0.9.0
django-extensions==3.2.1
django-filter==23.2
django-heroku==0.3.1
django-imagekit==2.0.4
django-ipware==4.0.2
django-jazzmin==2.5.0
django-js-asset==2.1.0
django-jsonview==2.0.0
django-map-widgets==0.4.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-oauth-toolkit==2.1.0
django-oauth2-provider==0.2.6.1
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-social-auth==0.7.1
django-storages==1.13.1
django-tinymce==3.5.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.4
duplicity==0.8.21
ecdsa==0.18.0
elastic-transport==8.4.0
elasticsearch==8.4.3
emoji==2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
exceptiongroup==1.1.1
executing==1.1.1
express==0.1.5
fasteners==0.14.1
filelock==3.8.0
Flask==2.2.2
flower==1.0.0
fonttools @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/fonttools-4.39.4-py3-none-any.whl#sha256=106caf6167c4597556b31a8d9175a3fdc0356fdcd70ab19973c3b0d4c893c461
future==0.18.2
GDAL==3.4.1
geographiclib==1.52
geopy==2.2.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
gpg==1.16.0
greenlet==1.1.3
h11==0.14.0
h2==4.1.0
heroku==0.1.4
hiredis==2.0.0
hpack==4.0.0
httplib2==0.22.0
humanize==4.6.0
hyperframe==6.0.1
hyperlink==21.0.0
idna @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/idna-3.4-py3-none-any.whl#sha256=90b77e79eaa3eba6de819a0c442c0b4ceefc341a7a2ab77d7562bf49f425c5c2
importlib-metadata==4.6.4
incremental==22.10.0
inflection==0.5.1
ipaddr==2.2.0
ipython==8.5.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonfield==3.1.0
jwcrypto==1.4.2
kaitaistruct==0.10
kazam==1.4.5
keyring==23.5.0
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
modeltranslation==0.25
monotonic==1.6
more-itertools==8.10.0
MouseInfo==0.1.3
mplcursors==0.5.2
mpld3 @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/mpld3-0.5.9-py3-none-any.whl#sha256=cecdd988b24f8d7c7f362b8f72bb91df1a24b1993dd02a958223918a31c31052
msgpack==1.0.5
msgpack-python==0.5.6
murmurhash==1.0.8
mysql-connector-python==8.0.31
natsort==8.0.2
netifaces==0.11.0
networkx==3.1
nltk==3.7
numpy==1.24.3
oauth2==1.9.0.post1
oauthlib==3.2.2
olefile==0.46
outcome==1.2.0
packaging==23.1
panda @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/panda-0.3.1.tar.gz#sha256=f213b848f09268b3e9fce0e103155ab003217c0e27f6048b6194e7f90bb2b716
pandas==2.0.2
pandas-datareader @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/pandas_datareader-0.10.0-py3-none-any.whl#sha256=0b95ff3635bc3ee1a6073521b557ab0e3c39d219f4a3b720b6b0bc6e8cdb4bb7
paramiko==2.9.3
parsel==1.7.0
parso==0.8.3
path==16.6.0
path.py==12.5.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.5.2
plotly @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/plotly-5.14.1-py2.py3-none-any.whl#sha256=a63f3ad9e4cc2e02902a738e5e3e7f3d1307f2732ac71a6c28f1238ed3052826
pluggy==0.13.0
preshed==3.0.7
prometheus-client==0.16.0
prompt-toolkit==3.0.33
Protego==0.2.1
protobuf==4.24.2
psycopg2==2.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.9.2
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.13.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyotp==2.6.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.28
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/python_dateutil-2.8.2-py2.py3-none-any.whl#sha256=961d03dc3453ebbc59dbdea9e4e11c5651520a876d0f4db161e8674aae935da9
python-debian==0.1.43+ubuntu1.1
python-decouple==3.6
python-dotenv==1.0.0
python-jose==3.3.0
python-openid==2.2.5
python-xlib==0.29
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==6.0
pyzmq==24.0.1
queuelib==1.6.2
ratelimit==2.2.1
rcssmin==1.1.0
redis==4.5.5
regex==2022.3.2
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rjsmin==1.2.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.5.2
Scrapy==2.8.0
SecretStorage==3.3.1
selenium==4.10.0
selenium-wire==5.1.0
service-identity==23.1.0
shortuuid==1.0.9
six @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/six-1.16.0-py2.py3-none-any.whl#sha256=8abb2f1d86890a2dfb989f9a77cfcfd3e47c2a354b01111771326f8aa26e0254
smart-open==5.2.1
smartsheet-python-sdk==3.0.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.41
sqlparse==0.4.4
srsly==2.4.4
stack-data==0.5.1
stripe==4.1.0
swapper==1.3.0
systemd-python==234
tenacity @ file:///home/studio45/Downloads/Filehandling13june/Filehandling/dependencies/tenacity-8.2.2-py3-none-any.whl#sha256=2f277afb21b851637e8f52e6a613ff08734c347dc19ade928e519d7d2d8569b0
textblob==0.17.1
thinc==8.1.4
tldextract==3.4.0
tornado==6.3.1
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.10.3
Twisted==23.8.0
txaio==23.1.1
typer==0.4.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
Unipath==1.1
uritemplate==4.1.1
urllib3==1.26.16
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.16.5
w3lib==2.1.1
wadllib==1.3.6
wasabi==0.10.1
watchdog==2.1.6
wcwidth==0.2.5
websockets==11.0.3
Werkzeug==2.2.2
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
zope.interface==6.0
zstandard==0.21.0
